+++ ether/cogs/utility/utils.py
@@ -14,7 +14,7 @@ from howlongtobeatpy import HowLongToBeat
 
 from ether.cogs.utility.uptime_cards import UptimeCards
 from ether.core.constants import Emoji
-from ether.core.embed import Embed, ErrorEmbed, SuccessEmbed
+from ether.core.embed import Embed
 from ether.core.i18n import _
 from ether.core.logging import log
 
@@ -29,9 +29,6 @@ def hltb_time(time: float) -> str:
 
     return f"{''.join(time)} Hours"
 
-print(__name__)
-
-
 class Utils(commands.GroupCog, name="utils"):
     def __init__(self, client):
         self.client = client
@@ -45,10 +42,10 @@ class Utils(commands.GroupCog, name="utils"):
         log.debug("Building monitor cards...")
         self.monitors_card = UptimeCards().card
 
-    @app_commands.command()
+    @app_commands.command(name="vote")
     @app_commands.checks.cooldown(1, 5.0, key=lambda i: (i.guild_id, i.user.id))
     async def vote(self, interaction: discord.Interaction) -> None:
-        """Get the link to vote for the bot"""
+        """Send a link to vote for Ether on top.gg."""
 
         await interaction.response.send_message(
             embed=Embed(
@@ -59,7 +56,7 @@ class Utils(commands.GroupCog, name="utils"):
             )
         )
 
-    @app_commands.command()
+    @app_commands.command(name="changelog")
     @app_commands.checks.cooldown(1, 5.0, key=lambda i: (i.guild_id, i.user.id))
     async def changelog(self, interaction: discord.Interaction) -> None:
         """Get the changelog"""
@@ -76,18 +73,13 @@ class Utils(commands.GroupCog, name="utils"):
 
         await interaction.response.send_message(f"```md\n{changelog}\n```")
 
-    @app_commands.command()
+    @app_commands.command(name="ping")
     @app_commands.checks.cooldown(1, 5.0, key=lambda i: (i.guild_id, i.user.id))
     async def ping(self, interaction: discord.Interaction) -> None:
         """Pong!"""
 
-        embed = SuccessEmbed(
-            title=":ping_pong:" + _("Pong!"),
-            description=_("Bot latency: `{x}ms`", x=round(self.client.latency * 1000)),
-        )
-
         await interaction.response.send_message(
-            embed=SuccessEmbed(
+            embed=Embed.success(
                 title=":ping_pong:" + _("Pong!"),
                 description=_(
                     "Bot latency: `{x}ms`", x=round(self.client.latency * 1000)
@@ -95,7 +87,7 @@ class Utils(commands.GroupCog, name="utils"):
             )
         )
 
-    @app_commands.command()
+    @app_commands.command(name="status")
     @app_commands.checks.cooldown(1, 5.0, key=lambda i: (i.guild_id, i.user.id))
     async def status(self, interaction: discord.Interaction) -> None:
         """Show the status of Ether"""
@@ -146,7 +138,7 @@ class Utils(commands.GroupCog, name="utils"):
         list = [i for i in items if i]
         return await interaction.response.send_message(choice(list))
 
-    @commands.command(name="roll")
+    @app_commands.command(name="roll")
     @app_commands.checks.cooldown(1, 5.0, key=lambda i: (i.guild_id, i.user.id))
     @app_commands.describe(dice="Dices to roll. (syntax: `1d6, 2d20+2`)")
     async def roll(
@@ -179,7 +171,7 @@ class Utils(commands.GroupCog, name="utils"):
 
         except ValueError:
             return await interaction.response.send_message(
-                embed=ErrorEmbed(
+                embed=Embed.error(
                     _("An error occurred, please check the syntax of your dices.")
                 ),
                 ephemeral=True,
@@ -243,7 +235,7 @@ class Utils(commands.GroupCog, name="utils"):
             await interaction.response.send_message(embed=embed)
         else:
             await interaction.response.send_message(
-                embed=ErrorEmbed('Could not find any definition of **"{term}"**!'),
+                embed=Embed.error('Could not find any definition of **"{term}"**!'),
                 delete_after=5,
             )
 
@@ -256,7 +248,7 @@ class Utils(commands.GroupCog, name="utils"):
             data = max(results_list, key=lambda element: element.similarity)
         else:
             return await interaction.response.send_message(
-                embed=ErrorEmbed("Sorry, we could not find your game."),
+                embed=Embed.error(description="Sorry, we could not find your game."),
                 ephemeral=True,
             )
 
@@ -317,7 +309,7 @@ class Utils(commands.GroupCog, name="utils"):
         timezone: Choice[str],
     ):
         """Get the next rocket launches"""
-        timezone = pytz.timezone(timezone)
+        timezone = pytz.timezone(timezone.value)
 
         r = requests.get("https://fdo.rocketlaunch.live/json/launches/next/5")
         res = r.json()
@@ -356,9 +348,10 @@ class Utils(commands.GroupCog, name="utils"):
     @app_commands.command(name="uptime")
     @app_commands.checks.cooldown(1, 15.0, key=lambda i: (i.guild_id, i.user.id))
     async def uptime(self, interaction: discord.Interaction):
+        """Get the uptime statistics of Ether"""
         if not self.monitors_card:
             return await interaction.response.send_message(
-                embed=ErrorEmbed(description="Sorry, uptime cards are not available yet, retry later.")
+                embed=Embed.error(description="Sorry, uptime cards are not available yet, retry later.")
             )
 
         await interaction.response.send_message(