+++ ether/core/tree.py
@@ -2,7 +2,7 @@ import discord
 import wavelink
 from discord import HTTPException, app_commands
 
-from ether.core.embed import ErrorEmbed
+from ether.core.embed import Embed
 from ether.core.logging import log
 
 
@@ -10,7 +10,7 @@ class Tree(app_commands.CommandTree):
     async def on_error(self, interaction: discord.Interaction, error: Exception):
         if isinstance(error, app_commands.CommandOnCooldown):
             return await interaction.response.send_message(
-                embed=ErrorEmbed(
+                embed=Embed.error(
                     description=f"This command is on cooldown, please retry in `{error.retry_after:.2f}s`."
                 ),
                 ephemeral=True,
@@ -30,12 +30,16 @@ class Tree(app_commands.CommandTree):
             return
 
         await interaction.response.send_message(
-            embed=ErrorEmbed(
+            embed=Embed.error(
                 description=f"An error occurred while executing this command, please retry later.\n If the problem persist, please contact the support.\n\n Error: `{error.__class__.__name__}({error})`"
             ),
             ephemeral=True,
         )
 
-        log.error(f"Error on command {interaction.command.name}")
-        log.error(f" => Selected parameters: {str(interaction.command.parameters)}")
+        parameters_display = []
+        if interaction.command:
+            for p in interaction.command.parameters: parameters_display.append(p.display_name)
+
+            log.error(f"Error on command {interaction.command.name}")
+            log.error(f" => Selected parameters: {', '.join(parameters_display)}")
         log.exception(error)